+++ partiture/src/main/kotlin/com/github/manosbatsis/partiture/flow/util/PartitureAccountsAwareFlow.kt
@@ -2,17 +2,20 @@ package com.github.manosbatsis.partiture.flow.util
 
 import co.paralleluniverse.fibers.Suspendable
 import com.github.manosbatsis.partiture.flow.PartitureFlow
+import com.github.manosbatsis.partiture.flow.call.CallContextEntry
 import com.github.manosbatsis.partiture.flow.io.input.InputConverter
 import com.github.manosbatsis.partiture.flow.io.output.OutputConverter
 import com.github.manosbatsis.partiture.flow.tx.initiating.SimpleTxStrategy
 import com.github.manosbatsis.partiture.flow.tx.initiating.TxStrategy
 import com.github.manosbatsis.partiture.flow.util.IdentitySyncMode.SKIP
+import com.r3.corda.lib.accounts.workflows.flows.ShareStateAndSyncAccounts
+import net.corda.core.contracts.ContractState
+import net.corda.core.flows.FlowException
 import net.corda.core.flows.FlowLogic
 import net.corda.core.flows.FlowSession
 import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.AnonymousParty
 import java.security.PublicKey
-
 
 
 /**
@@ -33,10 +36,22 @@ abstract class PartitureAccountsAwareFlow<IN, OUT>(
     override fun createFlowSessions(
             counterParties: Iterable<AbstractParty>
     ): Set<FlowSession> {
-        val parties = (counterParties)
+        val ourId = this.ourIdentity.name.toString()
+        println("$ourId createFlowSessions, parties: ${counterParties.joinToString (",") }}")
         val sessions = mutableSetOf<FlowSession>()
-        for(party in parties){
+        for(party in counterParties){
+            println("$ourId createFlowSessions, open session for AbstractParty: ${party}")
+
+            val wellKnown = toWellKnownParty(party)
+            println("$ourId createFlowSessions, open session for Party: ${wellKnown.name}")
+/*
+Our parties:     [Anonymous(DLEGd12j5kANU38Cgio9SfwxP4vDy9C99kwwAPyharm5GB),
+              Anonymous(DL34bbWXUpqJ4ULeXjhJVHw3GQkYmR8fSfUC8R1RQVznLB)] {actor_id=Only For Testing, actor_owning_identity=O=PartyB, L=New York, C=US, actor_store_id=TEST, fiber-id=10000009, flow-id=ecff83ae-d0b7-4d14-9658-1b35bd7cef0b, invocation_id=fdd4d7e7-1314-421d-abf8-4f5281145465, invocation_timestamp=2020-08-12T08:48:37.233Z, origin=Only For Testing, session_id=fdd4d7e7-1314-421d-abf8-4f5281145465, session_timestamp=2020-08-12T08:48:37.233Z, thread-id=1660, tx_id=A6AAC37BCFCDFDE01593CFD09E5917EF681B5E1F5387F2E4CB73B422C668BC29}
+Counter parties: [Anonymous(DLCHickUdb2oFss7HE2KnZt9s8cti2vbCQiLc62EvALAz6),
+              Anonymous(DLBC5NWkPosprW4kASn4BkGqi9M4vu4cvaHY98d925jqRt)] {actor_id=Only For Testing, actor_owning_identity=O=PartyB, L=New York, C=US, actor_store_id=TEST, fiber-id=10000009, flow-id=ecff83ae-d0b7-4d14-9658-1b35bd7cef0b, invocation_id=fdd4d7e7-1314-421d-abf8-4f5281145465, invocation_timestamp=2020-08-12T08:48:37.233Z, origin=Only For Testing, session_id=fdd4d7e7-1314-421d-abf8-4f5281145465, session_timestamp=2020-08-12T08:48:37.233Z,
+*/
             sessions.add(initiateFlow(party))
+
         }
         return sessions
     }
@@ -51,6 +66,23 @@ abstract class PartitureAccountsAwareFlow<IN, OUT>(
                 .map { it.owningKey }
         val accountKeys = ourParties.map { it.owningKey }
         return nodeKeys + accountKeys
-
     }
+
+
+    @Suspendable
+    override fun postExecuteFor(
+            ourParties: List<AbstractParty>,
+            counterParties: List<AbstractParty>,
+            sessions: Set<FlowSession>,
+            ccEntry: CallContextEntry) {
+        val tx = ccEntry.finalized ?: throw FlowException("CallContextEntry has no finalized TX")
+        val outputs = tx.coreTransaction.outRefsOfType<ContractState>()
+        for(state in outputs) {
+            for(session in sessions) {
+                // TODO: use ShareStateWithAccountFlow instead?
+                subFlow(ShareStateAndSyncAccounts(state, session.counterparty))
+            }
+        }
+    }
+
 }